While investigating HK2 classloader performance, I found that ResourceBundle.getBundle() is a rather expensive operation.
This is mainly because each load attempt turns into 3 invocations of loadClass (when it tries to load resource bundle class, a rarely used feature --- first with _en_US, then _en, finally the default locale), plus 3 invocations of getResource.

Everyone seems to grab a reference to StringManagers in a static field eagerly, so if StringManager tries to load getBundle() at the start-up, this adds a non-trivial pressure to classLoader, and thereby causing a visible delay to the start-up.

Those message resources are rarely used during the start up, so it stands to the reason to defer the loading. The lazy loading costs one comparison for every string look up, but that cost is negligible given the overall string format cost.

On my workstation, this change improved the start-up time by 10% when a simple war was deployed.

svn path=/trunk/; revision=19063