Partial fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=4727 ("session-destroyed-during-undeploy: FAIL"):
Fixed NPE that is thrown and swallowed when an attempt is made to enable SSO (this makes enabling SSO impossible).

This makes the above unit test pass, as long as the "asadmin set" command is commented out from its build.xml, and the virtual server's "sso-enabled" property is "manually" set to "true" in domain.xml.

Reviewed by: jerome

svn path=/trunk/; revision=19905