Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=4761 ("V3 Container Mapper broken for double context path, mapping redone by the web container")
Completely re-design the mapping algorithm grizzly-kernel was using, and instead use the Grizzly's Mapper class to map a request to its container, adapter and endpoint. The Mapper is now shared with the WebContainer so there is no duplicate mapping. When grizzly ARP or mod_jk is enabled, there is still double mapping as there is no way to share the Mapper as the invoking thread might not be the same as the one that created the Mapper.

Kudo to Jerome with his help on hk2/build issue.

+ ALL QL passed

(ready for rubber chicken now :-))


svn path=/trunk/; revision=21378


Former-commit-id: 40a0c43f5371af7ee801da93e2ad132d9f7414eb