Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=9690 ("[blocking]file-cache stats are not updated")
The fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=9286

re-enabled the optimization for skipping the double mapping of request|container, and exposed a new issue with the FileCache configuration. The complete fix will be in
tegrated with 1.9.18c for file located under the docroot, and the current commit fix the issue when the file is included in the application was itself.

+ ALL QL passed.

svn path=/trunk/; revision=32012


Former-commit-id: c4ab613a70f8fcb9daaa8964376076d9a15a479e